pyproject.toml
[tool.poetry]
name = "smart_meter_reader"    
version = "0.1.0"
description = ""
license = "MIT"
authors = ["Michael Hammer <mail@michael-hammer.at>"]

[tool.poetry.dependencies]
python = "^3.9"
pyserial = "^3.5"
dsmr-parser = "^0.33"
pycryptodome = "^3.15.0"
pyserial-asyncio = "^0.6"
aioserial = "^1.3.1"
dynaconf = "^3.1.9"
Homie4 = "^0.4.0"

[tool.poetry.dev-dependencies]
notebook = "^6.4.12"
jupyter = "^1.0.0"
black = "^22.6.0"
flake8 = "^5.0.4"
Sphinx = "^5.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
